backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-groq
python-dotenv
pdfminer.six
